requirements.txt
mysql-connector-python==8.1.0
pandas==2.0.3
python-dotenv==1.0.0